Move hashed() and authorize() into utils.js
